Saving / loading and NOI filter
Added apache common lang library to escape/unescape strings used in the
observation and action section of the tcc so they don't interfere with
reading of the xml document. Added a row filter to tables to filter out
rows that shouldn't be selected rather than altering the data behind the
table to remove rows.